requirements.txt
fastapi
uvicorn
python-barcode
segno
